Added user authentication and authorization

+++ api/src/controller/datasource.controller.ts
@@ -5,6 +5,9 @@ import { ApiOperationPost, ApiPath, SwaggerDefinitionConstant } from 'swagger-ex
 import { DataSourceService } from '../services/datasource.service';
 import { validate } from '../middleware/validation';
 import { DataSourceListRequest, DataSourceCreateRequest, DataSourceIDRequest } from '../api_models/datasource';
+import { RoleService } from '../services/role.service';
+import Account from '../models/account';
+import { ROLE_TYPES } from '../api_models/role';
 
 @ApiPath({
   path: '/datasource',
@@ -14,11 +17,14 @@ import { DataSourceListRequest, DataSourceCreateRequest, DataSourceIDRequest } f
 export class DataSourceController implements interfaces.Controller {
   public static TARGET_NAME: string = 'DataSource';
   private dataSourceService: DataSourceService;
+  private roleService: RoleService;
 
   public constructor(
-    @inject('Newable<DataSourceService>') DataSourceService: inverfaces.Newable<DataSourceService>
+    @inject('Newable<DataSourceService>') DataSourceService: inverfaces.Newable<DataSourceService>,
+    @inject('Newable<RoleService>') RoleService: inverfaces.Newable<RoleService>
   ) {
     this.dataSourceService = new DataSourceService();
+    this.roleService = new RoleService();
   }
 
   @ApiOperationPost({
@@ -35,6 +41,8 @@ export class DataSourceController implements interfaces.Controller {
   @httpPost('/list')
   public async list(req: express.Request, res: express.Response, next: express.NextFunction): Promise<void> {
     try {
+      const account: Account = req.body.account;
+      this.roleService.checkPermission(account, ROLE_TYPES.READER);
       const { filter, sort, pagination } = validate(DataSourceListRequest, req.body);
       const result = await this.dataSourceService.list(filter, sort, pagination);
       res.json(result);
@@ -57,6 +65,8 @@ export class DataSourceController implements interfaces.Controller {
   @httpPost('/create')
   public async create(req: express.Request, res: express.Response, next: express.NextFunction): Promise<void> {
     try {
+      const account: Account = req.body.account;
+      this.roleService.checkPermission(account, ROLE_TYPES.ADMIN);
       const { type, key, config } = validate(DataSourceCreateRequest, req.body);
       const result = await this.dataSourceService.create(type, key, config);
       res.json(result);
@@ -79,6 +89,8 @@ export class DataSourceController implements interfaces.Controller {
   @httpPost('/delete')
   public async delete(req: express.Request, res: express.Response, next: express.NextFunction): Promise<void> {
     try {
+      const account: Account = req.body.account;
+      this.roleService.checkPermission(account, ROLE_TYPES.ADMIN);
       const { id } = validate(DataSourceIDRequest, req.body);
       await this.dataSourceService.delete(id);
       res.json();
